requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
mysqlclient==2.2.0
SQLAlchemy>=2.0.21
Werkzeug==2.3.7
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
joblib==1.3.2
python-dotenv==1.0.0
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
urllib3==2.0.5
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
typing_extensions>=4.9.0
tzdata==2023.3
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
greenlet==2.0.2
scikit-learn>=1.0.0
xgboost>=1.7.0
lightgbm>=3.3.0
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.65.0
beautifulsoup4==4.12.2
html5lib==1.1
selenium>=4.18.1
webdriver-manager==4.0.1
flask-cors==4.0.0